requirements.txt
authorizenet==1.1.3
certifi==2019.11.28
cffi==1.14.4
chardet==3.0.4
defusedxml==0.7.0rc1
dj-database-url==0.5.0
Django==2.1.5
django-crispy-forms==1.7.2
django-heroku==0.3.1
django-localflavor==2.1
gunicorn==20.0.4
idna==2.8
lxml==4.5.0
oauthlib==3.1.0
psycopg2==2.7.3.2
pycparser==2.20
PyJWT==1.7.1
python3-openid==3.2.0
pytz==2018.9
PyXB==1.2.5
requests==2.22.0
requests-oauthlib==1.3.0
six==1.14.0
social-auth-app-django==4.0.0
#python-social-auth[django]
social-auth-core==3.3.3
twilio==6.35.4
urllib3==1.25.8
whitenoise==5.0.1
BeautifulSoup4
stripe==2.56.0